Refactor snapshotter entry point
