+++ snapshotter/cmd/naive/main.go
@@ -1,0 +1,33 @@
+// Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License"). You may
+// not use this file except in compliance with the License. A copy of the
+// License is located at
+//
+//	http://aws.amazon.com/apache2.0/
+//
+// or in the "license" file accompanying this file. This file is distributed
+// on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+// express or implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package main
+
+import (
+	"context"
+	"flag"
+
+	"github.com/containerd/containerd/snapshots"
+
+	"github.com/firecracker-microvm/firecracker-containerd/snapshotter"
+	"github.com/firecracker-microvm/firecracker-containerd/snapshotter/naive"
+)
+
+func main() {
+	var rootPath string
+	flag.StringVar(&rootPath, "path", "./images", "Path to snapshotter data (default: ./images)")
+
+	snapshotter.Run(func(ctx context.Context) (snapshots.Snapshotter, error) {
+		return naive.NewSnapshotter(ctx, rootPath)
+	})
+}
