--- snapshotter/cmd/snapshotter/.gitignore
@@ -1,1 +1,0 @@
-snapshotter
